frontend/web/requirements.txt
python-dotenv
Flask
spotipy
awswrangler
